requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
requests
httpx
python-dotenv
numpy
